Make sure right mount propagation is used
The rootfs and its parent mount need to be private for pivot_root;
however, we also want to enable users to add mounts and share them
across containers in the same pod.

To make sure pivot_root does not fail, we make the rootfs mount and its
parent mount (which is fixed to / right now) private right before
calling pivot_root, after a new mount namespace is created for the
container.