Simplify image name parsing
Image repos might have an arbitrary number of slashes, which breaks the
current heuristic in certain cases.

Instead, use a simple check to see if the first part of the image name
is a hostname, for example quay.io/myimage refers to the image repo
"myimage" on the registry server at quay.io.