Re-try growpart in resizeVolume()
Since we are calling /resizevolume a bit too early in Kip, growpart in
resizeVolume() might be started before the EBS volume does get resized.
I added a similar retry loop as with resize2fs to work around the issue.

Once we add a check of the status of the volume modification to Kip, I
assume this problem will go away, but IMO it's better to have in-depth
defense against it even once that check is in place.